Refactor image / video blocks to use duplicated media component
